refactor: use Logger for core debugging
