+++ packages/core/src/core/calendar-date.ts
@@ -10,6 +10,7 @@ import type {
 } from '../types/calendar';
 import { CalendarTimeUtils } from './calendar-time-utils';
 import { DateFormatter } from './date-formatter';
+import { Logger } from './logger';
 
 export class CalendarDate implements ICalendarDate {
   year: number;
@@ -44,7 +45,7 @@ export class CalendarDate implements ICalendarDate {
       this.calendar = calendar;
       this.formatter = new DateFormatter(calendar);
     } catch (error) {
-      console.debug('[S&S] Error creating CalendarDate:', error);
+      Logger.debug('Error creating CalendarDate:', error);
       throw error;
     }
   }
@@ -178,7 +179,7 @@ export class CalendarDate implements ICalendarDate {
         format: 'short',
       });
     } catch (error) {
-      console.debug('[S&S] Error formatting short date string:', error);
+      Logger.debug('Error formatting short date string:', error);
       return `${this.day} ${this.getMonthName('short')} ${this.getYearString()}`;
     }
   }
@@ -212,7 +213,7 @@ export class CalendarDate implements ICalendarDate {
         format: 'long',
       });
     } catch (error) {
-      console.debug('[S&S] Error formatting long date string:', error);
+      Logger.debug('Error formatting long date string:', error);
       const weekdayName = this.getWeekdayName('long');
       const monthName = this.getMonthName('long');
       const dayOrdinal = this.getDayString('long');
@@ -250,7 +251,7 @@ export class CalendarDate implements ICalendarDate {
         format: 'long',
       });
     } catch (error) {
-      console.debug('[S&S] Error formatting date string:', error);
+      Logger.debug('Error formatting date string:', error);
       const weekdayName = this.getWeekdayName('long');
       const monthName = this.getMonthName('long');
       const dayOrdinal = this.getDayString('long');
@@ -290,7 +291,7 @@ export class CalendarDate implements ICalendarDate {
     try {
       const weekday = this.calendar.weekdays?.[this.weekday];
       if (!weekday) {
-        console.debug(`[S&S] Invalid weekday index: ${this.weekday}`);
+        Logger.debug(`Invalid weekday index: ${this.weekday}`);
         return 'Unknown';
       }
 
@@ -300,7 +301,7 @@ export class CalendarDate implements ICalendarDate {
 
       return weekday.name || 'Unknown';
     } catch (error) {
-      console.debug('[S&S] Error getting weekday name:', error);
+      Logger.debug('Error getting weekday name:', error);
       return 'Unknown';
     }
   }
@@ -312,7 +313,7 @@ export class CalendarDate implements ICalendarDate {
     try {
       const month = this.calendar.months?.[this.month - 1];
       if (!month) {
-        console.debug(`[S&S] Invalid month index: ${this.month}`);
+        Logger.debug(`Invalid month index: ${this.month}`);
         return 'Unknown';
       }
 
@@ -322,7 +323,7 @@ export class CalendarDate implements ICalendarDate {
 
       return month.name || 'Unknown';
     } catch (error) {
-      console.debug('[S&S] Error getting month name:', error);
+      Logger.debug('Error getting month name:', error);
       return 'Unknown';
     }
   }
@@ -333,7 +334,7 @@ export class CalendarDate implements ICalendarDate {
   private getDayString(format: 'short' | 'long' | 'numeric'): string {
     try {
       if (typeof this.day !== 'number' || this.day < 1) {
-        console.debug(`[S&S] Invalid day value: ${this.day}`);
+        Logger.debug(`Invalid day value: ${this.day}`);
         return '1';
       }
 
@@ -348,7 +349,7 @@ export class CalendarDate implements ICalendarDate {
 
       return this.day.toString();
     } catch (error) {
-      console.debug('[S&S] Error formatting day string:', error);
+      Logger.debug('Error formatting day string:', error);
       return '1';
     }
   }
@@ -359,14 +360,14 @@ export class CalendarDate implements ICalendarDate {
   private getYearString(): string {
     try {
       if (typeof this.year !== 'number') {
-        console.debug(`[S&S] Invalid year value: ${this.year}`);
+        Logger.debug(`Invalid year value: ${this.year}`);
         return '1';
       }
 
       const { prefix = '', suffix = '' } = this.calendar.year || {};
       return `${prefix}${this.year}${suffix}`.trim();
     } catch (error) {
-      console.debug('[S&S] Error formatting year string:', error);
+      Logger.debug('Error formatting year string:', error);
       return this.year?.toString() || '1';
     }
   }
@@ -382,7 +383,7 @@ export class CalendarDate implements ICalendarDate {
 
       // Validate time components
       if (typeof hour !== 'number' || typeof minute !== 'number' || typeof second !== 'number') {
-        console.debug(`[S&S] Invalid time components:`, { hour, minute, second });
+        Logger.debug('Invalid time components:', { hour, minute, second });
         return '00:00:00';
       }
 
@@ -393,7 +394,7 @@ export class CalendarDate implements ICalendarDate {
 
       return `${hourStr}:${minuteStr}:${secondStr}`;
     } catch (error) {
-      console.debug('[S&S] Error formatting time string:', error);
+      Logger.debug('Error formatting time string:', error);
       return '00:00:00';
     }
   }
@@ -404,12 +405,12 @@ export class CalendarDate implements ICalendarDate {
   private addOrdinalSuffix(num: number): string {
     try {
       if (typeof num !== 'number' || num < 1) {
-        console.debug(`[S&S] Invalid number for ordinal suffix: ${num}`);
+        Logger.debug(`Invalid number for ordinal suffix: ${num}`);
         return '1st';
       }
       return CalendarTimeUtils.addOrdinalSuffix(num);
     } catch (error) {
-      console.debug('[S&S] Error adding ordinal suffix:', error);
+      Logger.debug('Error adding ordinal suffix:', error);
       return `${num || 1}th`;
     }
   }
@@ -437,7 +438,7 @@ export class CalendarDate implements ICalendarDate {
   compareTo(other: CalendarDateData): number {
     try {
       if (!other || typeof other !== 'object') {
-        console.debug('[S&S] Invalid date data provided for comparison:', other);
+        Logger.debug('Invalid date data provided for comparison:', other);
         return 0;
       }
 
@@ -454,7 +455,7 @@ export class CalendarDate implements ICalendarDate {
 
       return 0;
     } catch (error) {
-      console.debug('[S&S] Error comparing dates:', error);
+      Logger.debug('Error comparing dates:', error);
       return 0;
     }
   }
